Import of BDFOY/Apache-Htaccess-1.4 from CPAN.
gitpan-cpan-distribution: Apache-Htaccess
gitpan-cpan-version:      1.4
gitpan-cpan-path:         BDFOY/Apache-Htaccess-1.4.tar.gz
gitpan-cpan-author:       BDFOY
gitpan-cpan-maturity:     released

+++ t/parse.t
@@ -1,24 +1,35 @@
-# $Id: parse.t,v 1.1 2002/02/27 15:06:09 comdog Exp $
+# $Id: parse.t,v 1.3 2002/02/28 11:44:50 comdog Exp $
 
 BEGIN { $| = 1; print "1..3\n"; }
 END {print "not ok 1\n" unless $loaded;}
 use Apache::Htaccess;
-use File::Copy ();
-use Text::Diff;
+
 $loaded = 1;
 print "ok\n";
 
-my @test_files = qw( t/test.ht t/htaccess );
+if(  eval { require Text::Diff } and eval { require File::Copy } )
+	{
+	my @test_files = qw( t/test.ht t/htaccess );
+	
+	File::Copy::copy( @test_files );
+	
+	eval {
+		my $obj = Apache::Htaccess->new( $test_files[-1] );
+		print $obj ? '' : 'not ', "ok\n";
+		};
+	
+	my $diff = Text::Diff::diff( @test_files );
+	print $diff ? 'not ' : '', "ok\n";
+	if( $diff ) { print STDERR "\n$diff" };
 
-File::Copy::copy( @test_files );
-
-eval {
-	my $obj = Apache::Htaccess->new( $test_files[-1] );
-	print $obj ? '' : 'not ', "ok\n";
-	};
-
-my $diff = Text::Diff::diff( @test_files );
-print $diff ? 'not ' : '', "ok\n";
-if( $diff ) { print STDERR "\n$diff" };
-
-unlink $test_files[-1];
+	unlink $test_files[-1];
+	}
+else
+	{
+	my( $module ) = $@ =~ /Can't locate (.*?) in \@INC/;
+	
+	print STDERR "\nThe parsing test relies on Text::Diff and File::Copy\n" .
+		"but could not find one of those modules and so did not run.\n";
+	print "ok # Skip Could not load $1 (optional). Skipped test.\n";
+	print "ok # Skip Could not load $1 (optional). Skipped test.\n";
+	}
